Move map-related functions to map directory.
