+++ Makefile.in
@@ -129,8 +129,8 @@ TESTS    := $(addprefix $(TESTDIR)/,$(TESTS))
 OBJ      := main.o graphics.o events.o file.o
 OBJ      += util.o module.o optionparser.o parser.o state.o
 OBJ      += field/field.o 
-OBJ      += field/map.o field/mapview.o field/mapload.o
 OBJ      += field/object.o field/object-api.o field/object-image.o
+OBJ      += map/map.o map/mapview.o map/mapload.o
 
 
 # Note: DO NOT add .so or .dll onto the end of module names!
