+++ src/map/mapload.c
